+++ src/app/layout.tsx
@@ -16,6 +16,7 @@ import {
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
+  metadataBase: new URL(`${process.env.NEXT_PUBLIC_WEBSITE_URL}`),
   title: "Abdul Suleiman's Portfolio",
   description: "Software Developer",
   viewport: {
@@ -31,7 +32,7 @@ export const metadata = {
     creator: "@Abdul__Suleiman",
     title: "Abdul Suleiman's Portfolio",
     description: "Software Developer",
-    images: [`${process.env.NEXT_PUBLIC_METADATABASE_URL}card-bg.png`],
+    images: "/opengraph-image.png",
   },
   openGraph: {
     ...openGraphImage,
@@ -40,10 +41,12 @@ export const metadata = {
     icon: "/favicon.ico",
     shortcut: [...iconShortcut],
     apple: [...appleIcons],
-    other: {
-      rel: "apple-icon-precomposed",
-      url: "/favicon/apple-icon-precomposed.png",
-    },
+    other: [
+      {
+        rel: "apple-icon-precomposed",
+        url: "/favicon/apple-icon-precomposed.png",
+      },
+    ],
   },
   robots: {
     index: true,
