+++ src/app/opengraph-image.alt.txt
@@ -1,0 +1,1 @@
+opengraph-image
