+++ src/app/portfolio/project/[id]/page.tsx
@@ -17,7 +17,10 @@ export async function generateMetadata({ params }: ProjectProps) {
   if (docSnap?.exists()) {
     const { title } = docSnap?.data();
     return {
-      title: `Abdul Suleiman's Portfolio / ${title} `,
+      title: `Abdul Suleiman's Portfolio / ${title} → Project `,
+      alternates: {
+        canonical: `/project/${params?.id}`,
+      },
     };
   }
 }
