+++ src/app/robots.ts
@@ -1,0 +1,11 @@
+import { type MetadataRoute } from "next"
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: {
+      userAgent: "*",
+      allow: "/",
+    },
+    sitemap: `${process.env.NEXT_PUBLIC_METADATABASE_URL}sitemap.xml`,
+  }
+}
