+++ src/lib/utils.ts
@@ -4,4 +4,6 @@ import { twMerge } from 'tailwind-merge'
 
 export function cn(...inputs: ClassValue[]) {
     return twMerge(clsx(inputs))
-  }
+}
+
+
